storage/requirements.txt
asgiref==3.5.2
Django==4.1
django-cors-headers==3.13.0
django-filter==22.1
djangorestframework==3.13.1
gunicorn==20.0.4
python-dotenv==0.20.0
pytz==2022.1
sqlparse==0.4.2
